+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/StrikethroughMessageEntity.kt
@@ -1,0 +1,10 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity
+
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources.StrikethroughTextSource
+
+class StrikethroughMessageEntity(
+    override val offset: Int,
+    override val length: Int,
+    override val sourceString: String
+) : MessageEntity, TextSource by StrikethroughTextSource(sourceString)