+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/MessageEntity/textsources/UnderlineTextSource.kt
@@ -1,0 +1,12 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.textsources
+
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextSource
+import com.github.insanusmokrassar.TelegramBotAPI.utils.underlineHTML
+import com.github.insanusmokrassar.TelegramBotAPI.utils.underlineMarkdown
+
+class UnderlineTextSource(
+    sourceString: String
+) : TextSource {
+    override val asMarkdownSource: String = sourceString.underlineMarkdown()
+    override val asHtmlSource: String = sourceString.underlineHTML()
+}